refactor: move native tab to hotkey
